# internal/http/openapi/openapi.yaml
openapi: "3.1.0"
info:
  title: Suassu API
  description: API para gerenciamento de usuários e autenticação empresarial
  version: "1.0.0"
  contact:
    name: Suassu Team
    email: suporte@suassu.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

servers:
  - url: /api/v1
    description: API v1

paths:
  /healthz:
    get:
      summary: Health Check
      description: Endpoint para verificar o status da API
      operationId: healthCheck
      tags:
        - System
      responses:
        '200':
          description: API está funcionando
          content:
            text/plain:
              schema:
                type: string
                example: "OK"

  /auth/login:
    post:
      summary: Login de usuário
      description: Autentica um usuário e retorna um token JWT
      operationId: login
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignInInput'
      responses:
        '200':
          description: Login realizado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SignInOutput'
        '400':
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Credenciais inválidas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/validate-token:
    get:
      summary: Validar token JWT
      description: Valida um token JWT e retorna informações sobre sua validade
      operationId: validateToken
      tags:
        - Authentication
      parameters:
        - name: Authorization
          in: header
          required: true
          schema:
            type: string
            description: Token JWT para validação no formato "Bearer <token>"
            example: "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
      responses:
        '204':
          description: Token válido
        '200':
          description: Token inválido
          content:
            application/json:
              schema:
                type: object
                properties:
                  valid:
                    type: boolean
                    example: false
        '400':
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/me:
    get:
      summary: Informações do usuário autenticado
      description: Retorna informações do usuário baseado no token JWT
      operationId: getMe
      tags:
        - Authentication
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Informações do usuário retornadas com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserMeResponse'
        '401':
          description: Token inválido ou expirado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /users:
    post:
      summary: Criar usuário
      description: Cria um novo usuário na empresa
      operationId: createUser
      tags:
        - Users
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserInput'
      responses:
        '201':
          description: Usuário criado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateUserResponse'
        '400':
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Não autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      summary: Listar usuários
      description: Lista usuários da empresa com paginação por cursor
      operationId: listUsers
      tags:
        - Users
      security:
        - BearerAuth: []
      parameters:
        - name: limit
          in: query
          description: Número máximo de usuários por página
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 50
        - name: cursor
          in: query
          description: Cursor para paginação (base64 JSON com email e id)
          required: false
          schema:
            type: string
            example: "eyJlbWFpbCI6InVzdWFyaW9AZW1wcmVzYS5jb20iLCJpZCI6IjU1MGU4NDAwLWUyOWItNDFkNC1hNzE2LTQ0NjY1NTQ0MDAwMCJ9"
      responses:
        '200':
          description: Lista de usuários retornada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope'
        '401':
          description: Não autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Cursor inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtido através do endpoint /auth/login

  schemas:
    SignInInput:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: Email do usuário
          example: "usuario@empresa.com"
        password:
          type: string
          format: password
          description: Senha do usuário
          minLength: 1
          example: "senha123"

    SignInOutput:
      type: object
      required:
        - accessToken
      properties:
        accessToken:
          type: string
          description: Token JWT para autenticação
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."




    UserMeResponse:
      type: object
      required:
        - sub
        - email
        - name
        - enterpriseId
      properties:
        sub:
          type: string
          description: ID do usuário
          example: "550e8400-e29b-41d4-a716-446655440000"
        email:
          type: string
          format: email
          description: Email do usuário
          example: "usuario@empresa.com"
        name:
          type: string
          description: Nome completo do usuário
          example: "João Silva"
        enterpriseId:
          type: string
          description: ID da empresa
          example: "empresa-123"
        roleId:
          type: string
          nullable: true
          description: ID do papel/função do usuário
          example: "admin-role"

    CreateUserInput:
      type: object
      required:
        - name
        - email
        - password
        - document
      properties:
        name:
          type: string
          description: Nome completo do usuário
          minLength: 1
          example: "Maria Santos"
        email:
          type: string
          format: email
          description: Email do usuário
          example: "maria@empresa.com"
        password:
          type: string
          format: password
          description: Senha do usuário
          minLength: 1
          example: "senha123"
        document:
          type: string
          description: CPF ou CNPJ do usuário
          minLength: 1
          example: "123.456.789-00"
        phone:
          type: string
          nullable: true
          description: Telefone do usuário
          example: "+55 11 99999-9999"
        addressId:
          type: string
          nullable: true
          description: ID do endereço do usuário
          example: "addr-123"
        roleId:
          type: string
          nullable: true
          description: ID do papel/função do usuário
          example: "user-role"

    CreateUserResponse:
      type: object
      required:
        - id
      properties:
        id:
          type: string
          description: ID do usuário criado
          example: "550e8400-e29b-41d4-a716-446655440000"

    User:
      type: object
      required:
        - id
        - name
        - email
        - document
        - enterpriseId
      properties:
        id:
          type: string
          description: ID único do usuário
          example: "550e8400-e29b-41d4-a716-446655440000"
        name:
          type: string
          description: Nome completo do usuário
          example: "João Silva"
        email:
          type: string
          format: email
          description: Email do usuário
          example: "joao@empresa.com"
        document:
          type: string
          description: CPF ou CNPJ do usuário
          example: "123.456.789-00"
        phone:
          type: string
          nullable: true
          description: Telefone do usuário
          example: "+55 11 99999-9999"
        addressId:
          type: string
          nullable: true
          description: ID do endereço do usuário
          example: "addr-123"
        address:
          type: object
          properties:
            id:
              type: string
              example: "addr-123"
            zipCode:
              type: string
              example: "12345-678"
            state:
              type: string
              example: "SP"
            city:
              type: string
              example: "São Paulo"
            neighborhood:
              type: string
              example: "Centro"
            street:
              type: string
              example: "Avenida Paulista"
            num:
              type: string
              example: "1578"
            latitude:
              type: string
              example: "-23.5505"
            longitude:
              type: string
              example: "-46.6333"
            addInfo:
              type: string
              example: "Apto 123"
        roleId:
          type: string
          nullable: true
          description: ID do papel/função do usuário
          example: "user-role"
        enterpriseId:
          type: string
          description: ID da empresa
          example: "empresa-123"

    ErrorResponse:
      type: object
      required:
        - error
        - meta
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              enum: [invalid, not_found, conflict, unauthorized, forbidden, rate_limited, internal]
            message:
              type: string
            details:
              type: object
        meta:
          type: object
          required:
            - requestId
          properties:
            requestId:
              type: string

    Envelope:
      type: object
      required:
        - data
      properties:
        data:
          description: Dados da resposta
        meta:
          $ref: '#/components/schemas/MetaCursor'

    MetaCursor:
      type: object
      required:
        - limit
        - hasMore
      properties:
        limit:
          type: integer
          description: Número máximo de itens por página
        nextCursor:
          type: string
          description: Cursor para próxima página (base64 JSON)
        prevCursor:
          type: string
          description: Cursor para página anterior (base64 JSON)
        hasMore:
          type: boolean
          description: Indica se há mais páginas

tags:
  - name: System
    description: Endpoints do sistema
  - name: Authentication
    description: Endpoints de autenticação e autorização
  - name: Users
    description: Gerenciamento de usuários